Modal with styled-components - use GlobalStyles - transition KO on close
